Use hooks to build post- and term-field markup
The default markup for SEO fields for posts and terms isn't always
appropriate. For example, you might want to provide a "recommended
character count" in addition to (or instead of) the default character
count.

Until now, changing one part of the default markup has required
overriding a singleton method and replacing all of its content.

This commit adds hooks, admin template tags, and admin functions that
allow us to build the default markup incrementally and, in turn,
customize targeted pieces of it.

The admin template tags and functions also consolidate bits of logic and
markup that were previously duplicated across the edit-post, add-term,
and edit-term screens.

This commit also adds some general helpers that the new functions use:

* `wp_seo_intersect_args()`, which is `wp_parse_args()` without allowing
  you to add keys to the $defaults array.
* `WP_SEO::get_term_option()`, which gets the term meta option value
  for a term ID and taxonomy.
* `WP_SEO::intersect_term_option()`, which normalizes term meta
  option data via `wp_seo_intersect_args()`.