Handle $_POST slashing and meta API unslashing
To date, `WP_SEO::save_post_fields()` has not unslashed $_POST data before
processing it or slashed it before passing it to the metadata API (where
it is again unslashed).

In practice, unslashing and slashing the data is the same as just
leaving it. But unslashing $_POST data is a WordPress coding standard,
and the metadata API has always expected slashed values.

See also https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/172.