Move `wp_seo_get_key()` into `WP_SEO_Settings`
This also makes `$query` a required argument, rather than importing
the global `$wp_query`, which reduces the use of globals in the
plugin. The two calls to the method in `WP_SEO` already rely on
globals, so importing `$wp_query` to those is a wash.

Also, fix a few missed calls to global template tags rather than the
equivalent `WP_Query` methods.